#33 Playing with PUT /persist API
