+++ src/langium-model-server/lms/router.ts
@@ -1,12 +1,14 @@
-import * as http2 from 'http2'
-import { promisify } from 'util'
+import type * as http2 from 'http2'
 import { isPromise } from 'util/types'
-import type { SemanticIdentity } from '../identity/model'
 import type { IdentityIndex } from '../identity'
+import type { SemanticIdentity } from '../identity/model'
 import type { LangiumModelServerAddedServices, LmsServices } from '../services'
 import type { LmsDocument } from '../workspace/documents'
 import type { CreationParams, EditingResult } from './model'
 import { EditingFailureReason, Modification, Response, SemanticIdResponse } from './model'
+import { PathContainer } from './utils/path-container'
+import { readRequestBody, respondWithJson, setUpStreamForSSE } from './utils/http2-util'
+import * as vscode from 'vscode'
 
 type Http2RequestHandler = (stream: http2.ServerHttp2Stream, unmatchedPath: PathContainer,
     headers: http2.IncomingHttpHeaders, flags: number) => Http2RequestHandler | void
@@ -22,9 +24,7 @@ export function LangiumModelServerRouter<SM extends SemanticIdentity, II extends
         let handler: Http2RequestHandler
 
         if (unmatchedPath.hasPathSegments()) {
-            if (unmatchedPath.suffix === '') {
-                handler = helloWorldHandler
-            } else if (unmatchedPath.readPathSegment('models')) {
+            if (unmatchedPath.readPathSegment('models')) {
                 handler = provideModelHandler(services.lms)
             } else {
                 handler = notFoundHandler
@@ -46,14 +46,6 @@ export function LangiumModelServerRouter<SM extends SemanticIdentity, II extends
             }
         }
     }
-}
-
-const helloWorldHandler: Http2RequestHandler = (stream) => {
-    stream.respond({
-        'content-type': 'text/plain; charset=utf-8',
-        ':status': 200
-    })
-    stream.end('Hello World')
 }
 
 const provideModelHandler: Http2RequestHandlerProvider<LmsServices<SemanticIdentity>> = (sourceServices) => {
@@ -228,6 +220,19 @@ const provideModelHandler: Http2RequestHandlerProvider<LmsServices<SemanticIdent
             return
         }
 
+        const persistModelHandler: Http2RequestHandler = (stream, unmatchedPath) => {
+            const modelId = unmatchedPath.readPathSegment()
+            if (!modelId) {
+                return notFoundHandler
+            }
+            // workbench.action.files.saveLocalFile ?
+            vscode.commands.executeCommand('workbench.action.files.save').then((result) => {
+                console.debug('VSCode command executed with result', result)
+                respondWithJson(stream, Response.create('Executed Model Persist action', 200))
+            })
+            return
+        }
+
         const method = headers[':method']
         if (unmatchedPath.readPathSegment('subscriptions')) {
             if (method === 'POST')
@@ -237,6 +242,11 @@ const provideModelHandler: Http2RequestHandlerProvider<LmsServices<SemanticIdent
         if (unmatchedPath.readPathSegment('highlight')) {
             if (method === 'PUT')
                 return updateHighlightHandler
+            return notImplementedMethodHandler
+        }
+        if (unmatchedPath.readPathSegment('persist')) {
+            if (method === 'PUT')
+                return persistModelHandler
             return notImplementedMethodHandler
         }
         if (method === 'GET')
@@ -289,140 +299,4 @@ const notFoundHandler: Http2RequestHandler = (stream, unmatchedPath, header) => 
 const notImplementedMethodHandler: Http2RequestHandler = (stream, unmatchedPath, header) => {
     respondWithJson(stream,
         Response.create(`Path '${header[':path']}' is not implemented for method '${header[':method']}' (unmatched suffix '${unmatchedPath.suffix}')`, 501))
-}
-
-function readRequestBody(stream: http2.ServerHttp2Stream): Promise<object> {
-    const buffers: Array<Buffer | string> = []
-
-    stream.on('data', chunk => {
-        buffers.push(chunk)
-    })
-
-    return promisify(stream.once.bind(stream))('end')
-        .then(() => {
-            const joined = buffers.join()
-            console.debug('The stream ended. Joined first buffer', joined)
-            const parsed = JSON.parse(joined)
-            console.debug('Parsed response body', parsed)
-            return parsed
-        })
-}
-
-function respondWithJson(stream: http2.ServerHttp2Stream, response: Response): void
-function respondWithJson(stream: http2.ServerHttp2Stream, response: object, status: number): void
-function respondWithJson(stream: http2.ServerHttp2Stream, response: Response | object, status?: number): void {
-    console.debug('Responding with Response', response, status)
-    if (!status) {
-        status = (response as Response).code
-    }
-    stream.respond({
-        [http2.constants.HTTP2_HEADER_CONTENT_TYPE]: 'application/json; charset=utf-8',
-        [http2.constants.HTTP2_HEADER_STATUS]: status
-    })
-    if (status === 204) {
-        console.warn('You are trying to send response body with NO_CONTENT HTTP status. No body will be sent since the HTTP stream is already closed')
-        return
-    }
-    stream.end(JSON.stringify(response))
-}
-
-function setUpStreamForSSE(stream: http2.ServerHttp2Stream, response: Response): void
-function setUpStreamForSSE(stream: http2.ServerHttp2Stream, response: object, status: number): void
-function setUpStreamForSSE(stream: http2.ServerHttp2Stream, response: Response | object, status?: number): void {
-    if (!status) {
-        status = (response as Response).code
-    }
-    stream.respond({
-        [http2.constants.HTTP2_HEADER_STATUS]: status,
-        [http2.constants.HTTP2_HEADER_CONTENT_TYPE]: 'application/x-ndjson',
-        [http2.constants.HTTP2_HEADER_CACHE_CONTROL]: 'no-cache, no-store',
-    })
-    stream.write(JSON.stringify(response))
-}
-
-class PathContainer {
-    private _suffix: string
-
-    public get suffix(): string {
-        return this._suffix
-    }
-
-    public constructor(suffix: string) {
-        this._suffix = suffix
-    }
-
-    /**
-     * Checks whether unmatched path starts with '/'
-     */
-    public hasPathSegments(): boolean {
-        return (this.testPathSegmentSuffixMatch('') !== undefined)
-    }
-
-    /**
-     * Reads segment from the beginning of the path: `/${segmentValue}`. If specified `segmentValue` equals to the segment
-     * (i.e., it matches characters between the path delimimiter ('/') and another path delimiter or query character ('?') or end of the path),
-     * then segmentValue is returned and matched segment is eliminated from the {@link PathContainer}.
-     * If `segmentValue` is not provided, then attempts to match the path segment and return its value, also eliminating from the path.
-     * If the match is unsuccessful, returns `undefined` and leaves the path unchanged.
-     */
-    public readPathSegment(segmentValue?: string): string | undefined {
-        if (segmentValue !== undefined) {
-            if (segmentValue.length === 0) {
-                console.warn('Reading empty path segment (segmentValue is empty). This is most probably unintentionally')
-            }
-            const nextSegmentStart = this.testPathSegmentSuffixMatch(segmentValue)
-            if (nextSegmentStart === undefined || !this.isPathSegmentEnd(nextSegmentStart)) {
-                return undefined
-            }
-            this._suffix = this._suffix.substring(nextSegmentStart)
-            return segmentValue
-        }
-        return this.matchPrefix(/^\/[^\/?]*/)?.substring(1)
-    }
-
-    /**
-     * Reads query params (`?param=value&otherParam=value`) from the beginning of the path
-     */
-    public readQueryParams(): Record<string, string | undefined> | undefined {
-        const queryParamsText = this.matchPrefix(/^[?].*/)?.slice(1)
-        if (!queryParamsText) return undefined
-        const queryParams: Record<string, string | undefined> = {}
-        queryParamsText.split('&').forEach(queryParam => {
-            const assignment = queryParam.split('=', 2)
-            queryParams[assignment[0]] = assignment[1]
-        })
-        return queryParams
-    }
-
-    /**
-     * If successful, returns the index of the first unmatched character in the path after the matched path segment ('/' + segmentValue)
-     * Else returns `undefined`
-     */
-    private testPathSegmentSuffixMatch(segmentValue: string): number | undefined {
-        if (!this._suffix.startsWith('/' + segmentValue)) {
-            return undefined
-        }
-        return segmentValue.length + 1
-    }
-
-    private isPathSegmentEnd(pathIndex: number): boolean {
-        return this._suffix.length === pathIndex || this._suffix[pathIndex] === '/' || this._suffix[pathIndex] === '?'
-    }
-
-    /**
-     * If {@link PathContainer}.`suffix` begins with `prefix`, then this prefix is removed from `suffix`
-     * and method returns the matched prefix.
-     * Otherwise {@link PathContainer} remains unmodified and method returns `undefined`.
-     *
-     * @param prefix A RegExp against which existing `suffix` is matched
-     */
-    private matchPrefix(prefix: RegExp): string | undefined {
-        const matchResult = this._suffix.match(prefix)
-        if (!matchResult) {
-            return undefined
-        }
-        const prefixString = matchResult[0]
-        this._suffix = this._suffix.substring(prefixString.length)
-        return prefixString
-    }
 }
